Totally refactored hash logic
The hash module previously exposed a really inflexible interface.
Refactor the module to provide more useful operations. In doing so,
I discovered a mmap-leak in the ArchiveVolume versioning logic, and
also discovered that the hash checking logic didn't work at all
before the refactor. It appears to work now, since it took so much
effort to get it to.