Add lock-file feature for archive volumes:  * Fix a clunky feature of the FileHash interface (update calling    code accordingly)  * Add update-on-stale-lock member to ArchiveVolume YAML schema  * Add docker_volume_exists function to Docker interface  * Refactoring of the ArchiveVolume versioning logic slightly to support    new lock-file feature.
This feature will create a file with a ".lock" extension under
"/var/volumetric" for each volume that specifies the
"update-on-stale-lock" boolean as true in its YAML configuration. If
the modification time of the volume image is more recent than the
modification time of the lock file, and there is a mismatch between
the hash stored in the lock file and the hash of the configuration,
the volume will be updated, even if the volume already exists.